Move utils to vim_utils (plus a small polishing on funcref)
