+++ python3/vim_autoimport/vim_utils.py
@@ -1,0 +1,15 @@
+"""Vim-related utilities."""
+
+import vim
+import functools
+
+
+def funcref_nvim(name: str):
+    '''Wrap a vim function.'''
+    return functools.partial(vim.call, name)
+
+
+if hasattr(vim, 'Function'):
+    funcref = vim.Function
+else:
+    funcref = funcref_nvim
